+++ app/src/main/java/com/freshdigitable/yttt/YtttApp.kt
@@ -4,8 +4,4 @@ import android.app.Application
 import dagger.hilt.android.HiltAndroidApp
 
 @HiltAndroidApp
-class YtttApp : Application() {
-    val googleService: GoogleService by lazy {
-        GoogleService(this)
-    }
-}
+class YtttApp : Application()
